requirements.txt
python-dotenv==1.0.0
aiohttp==3.12.4
beautifulsoup4>=4.12.2
coloredlogs==15.0.1
python-telegram-bot==20.6
SQLAlchemy==2.0.21
pytz==2023.3.post1
aiofiles==23.2.1
nest_asyncio==1.5.8